WIP moved command errors to base router
