+++ robot-server/robot_server/runs/router/base_router.py
@@ -6,12 +6,16 @@ import logging
 from datetime import datetime
 from textwrap import dedent
 from typing import Optional, Union, Callable
-from typing_extensions import Literal
+from typing_extensions import Literal, Final
 
 from fastapi import APIRouter, Depends, status, Query
 from pydantic import BaseModel, Field
 
 from opentrons_shared_data.errors import ErrorCodes
+
+from opentrons.protocol_engine import (
+    errors as pe_errors,
+)
 
 from robot_server.errors.error_responses import ErrorDetails, ErrorBody
 from robot_server.protocols.protocol_models import ProtocolKind
@@ -55,6 +59,8 @@ from robot_server.service.notifications import get_pe_notify_publishers
 
 log = logging.getLogger(__name__)
 base_router = APIRouter()
+
+_DEFAULT_COMMAND_ERROR_LIST_LENGTH: Final = 20
 
 
 class RunNotFound(ErrorDetails):
@@ -402,5 +408,69 @@ async def put_error_recovery_policy(
 
     return await PydanticResponse.create(
         content=SimpleEmptyBody.construct(),
+        status_code=status.HTTP_200_OK,
+    )
+
+
+@PydanticResponse.wrap_route(
+    base_router.get,
+    path="/runs/{runId}/commandErrors",
+    summary="Get a list of all command errors in the run",
+    description=(
+        "Get a list of all command errors in the run. "
+        "\n\n"
+        "The errors are returned in order from oldest to newest."
+        "\n\n"
+        "This endpoint returns the command error. Use "
+        "`GET /runs/{runId}/commands/{commandId}` to get all "
+        "information available for a given command."
+    ),
+    responses={
+        status.HTTP_200_OK: {"model": MultiBody[list[pe_errors.ErrorOccurrence]]},
+        status.HTTP_409_CONFLICT: {"model": ErrorBody[RunStopped]},
+    },
+)
+async def get_run_commands_error(
+    runId: str,
+    cursor: Optional[int] = Query(
+        None,
+        description=(
+            "The starting index of the desired first command error in the list."
+            " If unspecified, a cursor will be selected automatically"
+            " based on the currently running or most recently executed command."
+        ),
+    ),
+    pageLength: int = Query(
+        _DEFAULT_COMMAND_ERROR_LIST_LENGTH,
+        description="The maximum number of command errors in the list to return.",
+    ),
+    run_data_manager: RunDataManager = Depends(get_run_data_manager),
+) -> PydanticResponse[MultiBody[list[pe_errors.ErrorOccurrence]]]:
+    """Get a summary of a set of command errors in a run.
+
+    Arguments:
+        runId: Requested run ID, from the URL
+        cursor: Cursor index for the collection response.
+        pageLength: Maximum number of items to return.
+        run_data_manager: Run data retrieval interface.
+    """
+    try:
+        command_error_slice = run_data_manager.get_command_error_slice(
+            run_id=runId,
+            cursor=cursor,
+            length=pageLength,
+        )
+    except RunNotCurrentError as e:
+        raise RunStopped(detail=str(e)).as_error(status.HTTP_409_CONFLICT) from e
+
+    meta = MultiBodyMeta(
+        cursor=command_error_slice.cursor,
+        totalLength=command_error_slice.total_length,
+    )
+
+    return await PydanticResponse.create(
+        content=MultiBody.construct(
+            data=command_error_slice.commands_errors, meta=meta
+        ),
         status_code=status.HTTP_200_OK,
     )
