+++ robot-server/tests/runs/router/test_commands_router.py
@@ -6,7 +6,6 @@ from decoy import Decoy, matchers
 
 from opentrons.protocol_engine import (
     CommandSlice,
-    CommandErrorSlice,
     CommandPointer,
     CommandNote,
     commands as pe_commands,
@@ -31,7 +30,6 @@ from robot_server.runs.router.commands_router import (
     get_run_command,
     get_run_commands,
     get_current_run_from_url,
-    get_run_commands_error,
 )
 
 
@@ -441,74 +439,6 @@ async def test_get_run_commands_not_found(
 
     assert exc_info.value.status_code == 404
     assert exc_info.value.content["errors"][0]["id"] == "RunNotFound"
-
-
-async def test_get_run_commands_errors(
-    decoy: Decoy, mock_run_data_manager: RunDataManager
-) -> None:
-    """It should return a list of all commands errors in a run."""
-    decoy.when(
-        mock_run_data_manager.get_command_error_slice(
-            run_id="run-id",
-            cursor=None,
-            length=42,
-        )
-    ).then_raise(RunNotCurrentError("oh no!"))
-
-    with pytest.raises(ApiError):
-        result = await get_run_commands_error(
-            runId="run-id",
-            run_data_manager=mock_run_data_manager,
-            cursor=None,
-            pageLength=42,
-        )
-        assert result.status_code == 409
-
-
-async def test_get_run_commands_errors_raises_no_run(
-    decoy: Decoy, mock_run_data_manager: RunDataManager
-) -> None:
-    """It should return a list of all commands errors in a run."""
-    error = (
-        pe_errors.ErrorOccurrence(
-            id="error-id",
-            errorType="PrettyBadError",
-            createdAt=datetime(year=2024, month=4, day=4),
-            detail="Things are not looking good.",
-        ),
-    )
-
-    command_error_slice = CommandErrorSlice(
-        cursor=1, total_length=3, commands_errors=list(error)
-    )
-
-    decoy.when(
-        mock_run_data_manager.get_command_error_slice(
-            run_id="run-id",
-            cursor=None,
-            length=42,
-        )
-    ).then_return(command_error_slice)
-
-    result = await get_run_commands_error(
-        runId="run-id",
-        run_data_manager=mock_run_data_manager,
-        cursor=None,
-        pageLength=42,
-    )
-
-    assert result.content.data == ResponseList(
-        __root__=[
-            pe_errors.ErrorOccurrence(
-                id="error-id",
-                errorType="PrettyBadError",
-                createdAt=datetime(year=2024, month=4, day=4),
-                detail="Things are not looking good.",
-            )
-        ]
-    )
-    assert result.content.meta == MultiBodyMeta(cursor=1, totalLength=3)
-    assert result.status_code == 200
 
 
 async def test_get_run_command_by_id(
